Big cleanup: - removed unused activities, fragments, menus - moved api.* to toplevel - ApiBottonActiviy is now MainActivity - moved earth/mars/system packages from view.api. to fragments. - repository.* is now nasa.*
