+++ app/src/main/java/com/geekbrains/potd/nasa/EpicPhotoDTO.kt
@@ -1,0 +1,9 @@
+package com.geekbrains.potd.nasa
+
+data class EpicPhotoDTO(
+    val identifier: String,
+    val caption: String,
+    val image: String,
+    val version: String,
+    val date: String,
+)
