+++ app/src/main/java/com/geekbrains/potd/nasa/EpicPhotosDTO.kt
@@ -1,0 +1,3 @@
+package com.geekbrains.potd.nasa
+
+typealias EpicPhotosDTO = ArrayList<EpicPhotoDTO>
