+++ app/src/main/java/com/geekbrains/potd/nasa/MarsPhotosDTO.kt
@@ -1,0 +1,7 @@
+package com.geekbrains.potd.nasa
+
+import com.google.gson.annotations.SerializedName
+
+data class MarsPhotosDTO(
+    @field:SerializedName("photos") val photos: ArrayList<MarsPhotoDTO>,
+)
