+++ app/src/main/java/com/geekbrains/potd/nasa/NasaRetrofitImpl.kt
@@ -1,0 +1,15 @@
+package com.geekbrains.potd.nasa
+
+import com.google.gson.GsonBuilder
+import retrofit2.Retrofit
+import retrofit2.converter.gson.GsonConverterFactory
+
+class NasaRetrofitImpl {
+    val nasaApi: NasaApi by lazy {
+        val retrofit = Retrofit.Builder()
+            .baseUrl("https://api.nasa.gov/")
+            .addConverterFactory(GsonConverterFactory.create(GsonBuilder().setLenient().create()))
+            .build()
+        retrofit.create(NasaApi::class.java)
+    }
+}
