+++ app/src/main/java/com/geekbrains/potd/nasa/PotdDTO.kt
@@ -1,0 +1,14 @@
+package com.geekbrains.potd.nasa
+
+import com.google.gson.annotations.SerializedName
+
+data class PotdDTO (
+    @field:SerializedName("copyright") val copyright: String?,
+    @field:SerializedName("date") val date: String?,
+    @field:SerializedName("explanation") val explanation: String?,
+    @field:SerializedName("media_type") val mediaType: String?,
+    @field:SerializedName("title") val title: String?,
+    @field:SerializedName("url") val url: String?,
+    @field:SerializedName("hdurl") val hdurl: String?,
+)
+
