--- app/src/main/java/com/geekbrains/potd/repository/EpicPhotoDTO.kt
@@ -1,9 +1,0 @@
-package com.geekbrains.potd.repository
-
-data class EpicPhotoDTO(
-    val identifier: String,
-    val caption: String,
-    val image: String,
-    val version: String,
-    val date: String,
-)
