--- app/src/main/java/com/geekbrains/potd/repository/EpicPhotosDTO.kt
@@ -1,3 +1,0 @@
-package com.geekbrains.potd.repository
-
-typealias EpicPhotosDTO = ArrayList<EpicPhotoDTO>
