--- app/src/main/java/com/geekbrains/potd/repository/MarsCameraDTO.kt
@@ -1,8 +1,0 @@
-package com.geekbrains.potd.repository
-
-import com.google.gson.annotations.SerializedName
-
-data class MarsCameraDTO(
-    @field:SerializedName("full_name") val fullName: String?,
-    @field:SerializedName("name") val shortName: String?,
-)
