--- app/src/main/java/com/geekbrains/potd/repository/MarsPhotoDTO.kt
@@ -1,9 +1,0 @@
-package com.geekbrains.potd.repository
-
-import com.google.gson.annotations.SerializedName
-
-data class MarsPhotoDTO(
-    @field:SerializedName("img_src") val url: String?,
-    @field:SerializedName("earth_date") val date: String?,
-    @field:SerializedName("camera") val camera: MarsCameraDTO?,
-)
