--- app/src/main/java/com/geekbrains/potd/repository/MarsPhotosDTO.kt
@@ -1,7 +1,0 @@
-package com.geekbrains.potd.repository
-
-import com.google.gson.annotations.SerializedName
-
-data class MarsPhotosDTO(
-    @field:SerializedName("photos") val photos: ArrayList<MarsPhotoDTO>,
-)
