--- app/src/main/java/com/geekbrains/potd/repository/NasaRetrofitImpl.kt
@@ -1,15 +1,0 @@
-package com.geekbrains.potd.repository
-
-import com.google.gson.GsonBuilder
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-
-class NasaRetrofitImpl {
-    val nasaApi: NasaApi by lazy {
-        val retrofit = Retrofit.Builder()
-            .baseUrl("https://api.nasa.gov/")
-            .addConverterFactory(GsonConverterFactory.create(GsonBuilder().setLenient().create()))
-            .build()
-        retrofit.create(NasaApi::class.java)
-    }
-}
