--- app/src/main/java/com/geekbrains/potd/view/ChangeThemeCallback.kt
@@ -1,5 +1,0 @@
-package com.geekbrains.potd.view
-
-interface ChangeThemeCallback {
-    fun onChangeTheme(themeId: Int)
-}
