--- app/src/main/java/com/geekbrains/potd/view/picture/BottomNavFragment.kt
@@ -1,42 +1,0 @@
-package com.geekbrains.potd.view.picture
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import android.widget.Toast
-import com.geekbrains.potd.R
-import com.geekbrains.potd.databinding.BottomNavigationLayoutBinding
-import com.google.android.material.bottomsheet.BottomSheetDialogFragment
-
-class BottomNavFragment : BottomSheetDialogFragment() {
-    private var _binding: BottomNavigationLayoutBinding? = null
-    val binding: BottomNavigationLayoutBinding get() = _binding!!
-
-    override fun onDestroy() {
-        super.onDestroy()
-        _binding = null
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-        binding.navigationView.setNavigationItemSelectedListener { menuItem ->
-            when(menuItem.itemId) {
-                R.id.navigation_one -> Toast.makeText(context, "One!", Toast.LENGTH_SHORT).show()
-                R.id.navigation_two -> Toast.makeText(context, "Two!", Toast.LENGTH_SHORT).show()
-                R.id.navigation_tri -> Toast.makeText(context, "Tri!", Toast.LENGTH_SHORT).show()
-            }
-            true
-        }
-    }
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        _binding = BottomNavigationLayoutBinding.inflate(inflater, container, false)
-        return binding.root
-    }
-
-}
