--- app/src/main/java/com/geekbrains/potd/viewmodel/POTDState.kt
@@ -1,9 +1,0 @@
-package com.geekbrains.potd.viewmodel
-
-import com.geekbrains.potd.repository.PotdDTO
-
-sealed class POTDState {
-    data class Success(val potdResponse: PotdDTO): POTDState()
-    data class Loading(val progress: Int?): POTDState()
-    data class Error(val error: Throwable): POTDState()
-}
