+++ app/src/main/res/layout/fragment_system.xml
@@ -44,7 +44,7 @@
 
     </com.google.android.material.chip.ChipGroup>
 
-    <com.geekbrains.potd.utils.CustomImageView
+    <ImageView
         android:id="@+id/imageView"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
