--- app/src/main/res/menu/menu_bottom_bar.xml
@@ -1,28 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-
-    <item
-        android:id="@+id/app_bar_theme"
-        android:icon="@drawable/ic_eye"
-        android:title="@string/theme"
-        app:showAsAction="ifRoom">
-        <menu>
-            <item
-                android:id="@+id/theme_steel"
-                android:title="@string/theme_steel_desc"
-                app:showAsAction="ifRoom" />
-            <item
-                android:id="@+id/theme_copper"
-                android:title="@string/theme_copper_desc"
-                app:showAsAction="ifRoom" />
-        </menu>
-    </item>
-
-    <item
-        android:id="@+id/app_bar_settings"
-        android:icon="@drawable/ic_settings_menu"
-        android:title="@string/settings"
-        app:showAsAction="ifRoom" />
-
-</menu>
