--- app/src/main/res/menu/menu_bottom_navigation.xml
@@ -1,19 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <item
-        android:id="@+id/navigation_one"
-        android:icon="@drawable/ic_archive"
-        android:title="На экран 1" />
-
-    <item
-        android:id="@+id/navigation_two"
-        android:icon="@drawable/ic_send"
-        android:title="На экран 2" />
-
-    <item
-        android:id="@+id/navigation_tri"
-        android:icon="@drawable/ic_search"
-        android:title="На экран 3" />
-
-</menu>
